requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cmdstanpy==1.2.4
colorama==0.4.6
contourpy==1.3.3
curl_cffi==0.11.4
cycler==0.12.1
dateparser==1.2.2
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
finnhub-python==2.4.24
fonttools==4.59.2
frozendict==2.4.6
frozenlist==1.7.0
holidays==0.79
idna==3.10
importlib_resources==6.5.2
ipython==8.12.3
jedi==0.19.2
joblib==1.5.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
mistune==3.1.3
multidict==6.6.0
multitasking==0.0.11
nbclient==0.10.2
nbformat==5.10.4
numpy==2.2.6
packaging==25.0
pandas==2.3.2
pandas_ta==0.4.71b0
pandocfilters==1.5.1
parso==0.8.4
peewee==3.18.1
pickleshare==0.7.5
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==1.7.1
pyparsing==3.2.3
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rest-framework-simplejwt==0.0.2
rpds-py==0.26.0
scikit-learn==1.6.0
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
stanio==0.5.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
yarg==0.1.9
yarl==1.20.1
yfinance==0.2.55
gunicorn
whitenoise
psycopg[binary]
dj-database-url
pywin32==311; platform_system == "Windows"

